+++ src/main/java/fr/istic/vv/operations/Plus.java
@@ -12,7 +12,7 @@ public class Plus implements Operation {
      * @throws OperationException
      */
     @Override
-    public double execute(int op1, int op2) throws OperationException {
+    public double execute(double op1, double op2) throws OperationException {
         return op1 + op2;
     }
 }
